Bugfix for get_pseudo_wave_function in ASE interface - dtype was moved from paw to wfs.
git-svn-id: https://svn.fysik.dtu.dk/projects/gpaw/trunk@3389 6419b465-d902-0410-997c-be2266baf723

+++ gpaw/aseinterface.py
@@ -332,7 +332,7 @@ class GPAW(PAW):
         psit_G = self.wfs.get_wave_function_array(band, kpt, spin)
         if broadcast:
             if not self.wfs.world.rank == 0:
-                psit_G = self.gd.empty(dtype=self.dtype, global_array=True)
+                psit_G = self.gd.empty(dtype=self.wfs.dtype, global_array=True)
             self.wfs.world.broadcast(psit_G, 0)
             return psit_G / Bohr**1.5
         elif self.wfs.world.rank == 0:
